Refactor based on PR comments
I removed the qualifier tags and bean names. Only left a couple that seemed to be required in test scripts.
Also removed the extend MaatApiClient to contribution and FDC clients, even though, the intention was to bind that the Client Factpry will always use a client member of MaatApiClient family.. but that doesn't seem to be working